move the TruncateFor helpers into the sqlparser package
