+++ go/sqldb/sql_error.go
@@ -9,6 +9,8 @@ import (
 	"fmt"
 	"regexp"
 	"strconv"
+
+	"github.com/youtube/vitess/go/vt/sqlparser"
 )
 
 const (
@@ -49,7 +51,7 @@ func (se *SQLError) Error() string {
 	fmt.Fprintf(buf, " (errno %v) (sqlstate %v)", se.Num, se.State)
 
 	if se.Query != "" {
-		fmt.Fprintf(buf, " during query: %s", TruncateForError(se.Query))
+		fmt.Fprintf(buf, " during query: %s", sqlparser.TruncateForError(se.Query))
 	}
 
 	return buf.String()
