+++ go/vt/sqlparser/parsed_query.go
@@ -10,7 +10,6 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/youtube/vitess/go/sqldb"
 	"github.com/youtube/vitess/go/sqltypes"
 
 	querypb "github.com/youtube/vitess/go/vt/proto/query"
@@ -54,7 +53,7 @@ func (pq *ParsedQuery) GenerateQuery(bindVariables map[string]interface{}) ([]by
 // MarshalJSON is a custom JSON marshaler for ParsedQuery.
 // Note that any queries longer that 512 bytes will be truncated.
 func (pq *ParsedQuery) MarshalJSON() ([]byte, error) {
-	return json.Marshal(sqldb.TruncateForUI(pq.Query))
+	return json.Marshal(TruncateForUI(pq.Query))
 }
 
 // EncodeValue encodes one bind variable value into the query.
