+++ go/vt/sqlparser/truncate_query.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package sqldb
+package sqlparser
 
 import (
 	"flag"
