+++ go/vt/vtgate/planner.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/youtube/vitess/go/acl"
 	"github.com/youtube/vitess/go/cache"
-	"github.com/youtube/vitess/go/sqldb"
 	"github.com/youtube/vitess/go/vt/sqlparser"
 	"github.com/youtube/vitess/go/vt/topo"
 	"github.com/youtube/vitess/go/vt/vtgate/engine"
@@ -336,7 +335,7 @@ func (plr *Planner) ServeHTTP(response http.ResponseWriter, request *http.Reques
 		response.Header().Set("Content-Type", "text/plain")
 		response.Write([]byte(fmt.Sprintf("Length: %d\n", len(keys))))
 		for _, v := range keys {
-			response.Write([]byte(fmt.Sprintf("%#v\n", sqldb.TruncateForUI(v))))
+			response.Write([]byte(fmt.Sprintf("%#v\n", sqlparser.TruncateForUI(v))))
 			if plan, ok := plr.plans.Peek(v); ok {
 				if b, err := json.MarshalIndent(plan, "", "  "); err != nil {
 					response.Write([]byte(err.Error()))
